Updated tests

+++ src/api/test/smoke.test.ts
@@ -3,7 +3,7 @@ import { generateNewProduct } from "../../data/products/productGeneration.js";
 // import productService from "../services/product.service.js";
 
 describe("Product Api tests", () => {
-  let product: Product;
+  let product: Product | null;
   beforeEach(async () => {});
 
   it("Should create product", async () => {
@@ -12,7 +12,17 @@ describe("Product Api tests", () => {
     expect(product.getProductFieldsFromSettings()).toMatchObject({ ...productData });
   });
 
+  it("Should update product", async () => {
+    const productData = generateNewProduct();
+    product = await Product.create(productData);
+    await product.edit({ ...productData, _id: product.getSettings()._id });
+    expect(product.getProductFieldsFromSettings()).toMatchObject({ ...productData });
+  });
+
   afterEach(async () => {
-    await product.delete();
+    if (product) {
+      await product.delete();
+    }
+    product = null;
   });
 });
